Expose a nicer getSourceFolders API in the variant.
The existing API (getSourceSets) does not include the generated
source folders and therefore is of limited usage for someone wanting
to consume the source folders. It also has other limitations like
not supporting includes/excludes pattern that are really needed
at the build level (even if the IDE doesn't fully support it yet.)

The new API returns a list of ConfigurableFileTree that includes the root
folder, the includes/excludes patterns, and the task dependency that
generated the source content (if applicable).

Change-Id: I99fb2c260c3817590d175f9a81b42c1886c71923
Test: existing.