Don't use converted kt templates when using O r2
When a developer chooses O r2, projects could still be using
26.0.0-alpha1 of the appcompat library which uses the older
findViewById(). This causes compilation issues in the new project.

Instead of using converted templates, rely on the runtime code
conversion from the Extension point when buildVersion == 26
and revision != 1. It looks up the signature before converting
code ensuring that the resulting code compiles.

Test: manual
Change-Id: I0302a1b36f0c32664863e1b58dc494a4086f020a