Fix studio freezing due to concurrency issue
We were locking buffer_mutex_ in NetworkCollector::Stop scope and then
calling profiler_thread_.join() from the same scope. The problem is
profiler_thread_::join can indirectly call StoreDataToBuffer(), which
uses the same mutex. In this scenario, which could be reproduced
sometimes by changing devices/processes, Studio was freezing.

This CL changes the threading logic so the thread is only created once
(in the constructor) and joined once (in the destructor), eliminating
the chance for the thread and the Start/Stop functions from
deadlocking each other.

Test: perfd network tests still pass. Manual testing was done as well.
Bug: 38350467
Bug: 38341792
Bug: 38357075
Change-Id: Iaf4d5efcf5dd8d621f25d317a0123daf70a53033