Fix OsProcessRunnerTest to accomdate change in adb
The test relied on adb printing out the help message on stderr. That is
not true for adb 1.0.39+. This CL fixes the test to accomodate both
cases. We need to support both cases since temporarily, our repo
checkout has different versions of adb in different platforms (windows
is on the latest, whereas linux and mac are on the older version)

Test: updated
Change-Id: I86748e0e09d3142311e0db2fde21c7d743cffadd