Add a default value for instantAppActivityRouteType
This CL adds a default value for the
instantAppActivityRouteType. This is required as it is
possible to create a template that uses the instant app
activity templates without defining the value in
template.xml. This creates a situation where it is
possible for the template to be attempt to access an
uninitialised variable which would cause it to fail.

Test: TemplateTest
Change-Id: Ic1cdfa948d4589f0e479ef7a0615713390b90760