Publish full jar as intermediate artifact.
When artifactType of 'jar' is requested from a library project, it was
returning a jar containg just the classes.  This change create a task
that merges classes.jar and res.jar into full.jar.  The task is
normally not run unless 'jar' is requested from a Configuration
artifactView.

Also simplify usage of HelloWorldLibraryApp.

Test: ZipMergeingTaskTest and LibraryIntermediateArtifactPublishingTest
Change-Id: I1e176058d4040928dbfbe8c9af5321a7d64851f0