Fix Proxy#invoke throwing UndeclaredThrowableException
In an actual app, an OkHttp interceptor interrupted its chain by
throwing a closed-socket related IOException. While OkHttp is
designed to recover gracefully from this, our reflection code
instead ended up wrapping the exception in a
TargetInvocationException, which OkHttp was not prepared for.

We fix this by unwrapping the TargetInvocationException, exposing
the original exception underneath.

Bug: 38490207
Test: N/A
Change-Id: I664036d9d1cc24e69c3b026c3179115624d2bcbb