[IAPP] AAPT2 link failed while running Master/Detail Flow template
Strings used on the base manifest need to be on the base fragment. At
the moment only "app_name" was needed, but "Master Detail Flow" needs
one more.
Use "baseFeatureResOut" template variable as the destination of the
extra strings (strings_iapp.xml.ftl). If we are building an instant app,
this variable points to the base resource, if we are building a normal
app, the variable points to the "app" resource location.

Test: Covered by existing template tests and UI tests
Change-Id: Iabef67a9ba48c1f7468095b035fb43b307a425d8