devicelib: Remove unnecessary CommandBuffer abstraction
All the uses of the CommandBuffer looked like:

  buffer = new CommandBuffer().writeHostCommand(cmd)
  result = conn.executeCommand(buffer)

The idea behind that was to have the command buffer be in charge of
formatting arguments to commands. But the formatting varies with each
command, so it makes sense for that to stay within each command itself.

This CL simplifies the codebase by deleting the command buffer and
instead have a connection be able to directly execute an AdbCommand
instance, each of which knows how to format itself.

In addition, this CL also deletes the HostService enum, inlining its
constants into each AdbCommand.

Test: existing tests pass
Change-Id: Ib85cc4b2c46430fd77ef8fa50df28d814788e62a