New DSL to specify alternates for build types and flavors.
The DSL looks like this:
android {
  buildTypeMatching 'staging', 'debug'
  productFlavorMatching 'color', 'blue', 'cyan'
}

The first item makes a project with 'staging' able
to select 'debug' instead in a dependency (if staging
does not exsist. If it exists 'staging' will be selected).

The 2nd example does the same for the flavor 'blue',
in dimension 'color', and adds 'cyan' as an alternate.
Again if 'blue' is present in the consumed project, then it
is used.

Bug: 62241369
Test: new integration test: AttributeMatchingTest
Change-Id: Ib1a568a34c9a4692c5fa233f40cd6d4e70e5a084