Add support for lint command line driver to be configured via XML file
Until now the lint command line driver could either

(1) read Eclipse-based projects (e.g. with src, res,
    AndroidManifest.xml layout and optionally project.properties and
    .classpath files configuring dependencies and compileSdkVersion)

(2) analyze a "flattened" project where you point to the sources,
    resources and manifests manually via --sources, --resources, and
    so on.

This CL adds support for a third option, --project <file.xml>,
where the given project descriptor file contains a definition of
the project and module layout and the dependencies between them.

This lets other build systems run lint on the source code without
having to present a flattened view of the world (which breaks certain
rules that consider project layout, such as rules that only apply in
library projects and so on.)

Test: Tests included

Change-Id: I37b885f0f5aa50e0937b114192642cb43f3aad4f