Fix Peripheral template for sensors & inputs
Services for sensors need to be declared in the manifest. Otherwise
they won't be run when the main activity starts the intent. Moreover,
Android Things now requires new permissions to use sensors & inputs.
So this template now adds those permissions in the manifest.

Test: Fix Peripheral template for sensors & inputs

Services for sensors need to be declared in the manifest. Otherwise
they won't be run when the main activity starts the intent. Moreover,
Android Things now requires new permissions to use sensors & inputs.
So this template now adds those permissions in the manifest.

Test: TemplateTest passes and manually tested at b/64578980
Bug: 64576913

Change-Id: I8f94169678dec1c358c79e8c040f5a87df0c80ea