Improve Jacoco maven version fetching
This CL uses the plugin's classloader
to get the Jacoco maven version that is used
to instrument classes. In addition to that,
resolution strategy is added that forces
the agent runtime to be the same version as
used for offline instrumentation.

Compared to the original CL,
ag/Ieb5740b134dffb477042796a55f68f9344430164, this
CL is using maven properties file from the jar as
using JaCoCo.VERSION did not provide the maven version.

BUG=37116868
Test: JacocoDependenciesTest

Change-Id: I1527e8a2a2ce0a92658a483aeec5c4206398c598