Fix deprecated Guava usages in FileCache.
FileCache now uses Hashing.sha256() as the hashing function for
computing cache keys. The new function is recommended by Guava as a
replacement for the deprecated Hashing.sha1() function.

With this change, cache entries created in Android Gradle plugin 3.0
will not be shared with those created in 2.3.x and earlier.

However, even without this change, that is already the case, as the
cache key computation is already different in 3.0 compared to 2.3.x.

Therefore, it is better for this change to go into the 3.0 release
rather than 3.1. (If introduced in 3.1, it would make the build cache
grow larger one more time unnecessarily.)

Test: Updated FileCacheTest
Change-Id: Ida9a3bda04239887014ca3e3b360a918d964cad9