Make AndroidSourceDirectorySet.getSrcDirs return ImmutableSet.
We've seen some developers try to write:
android.sourceSets.jni.srcDirs += ...

However because the returned Set was a new object allocated
just for this, the += operation did nothing to the internal
storage.

Returning an immutable collection solves the problem by
triggering a proper error when trying to add items to it.

Bug: 64436431
Test: existing.
Change-Id: Ib8241c27d8c3faf79cc2a61f1938a697f3075fb4