Remove search for a specific CMake version
Issue: Currently we search for a specific version of CMake (3.6.4111459,
which is the latest release). So if the user has an older version of
CMake, then we wouldn't be able to find it and build would fail.

Fix: do not search for the specific version

Test: 1. integration tests
2. Create a sample app and set the SDK with older version of CMake and
build it and run the app.

Bug: 65494776

cmake-version-fix

Change-Id: Ie9476a2a7d8e7fd4dce50d4fba8f5865431d2c01