Manually add renderscript compat to test variant.
It used to be added automatically since the test variant
manually got the classpath of the tested variant.

Now that it's done via configuration extension, this stopped working
because it's added the resolve configuration instead of the intermediate
(per-flavor/type) ones where the extension happens.

The better fix would be to properly setup these dependencies but
this is too complex and will have to wait after 3.0.
For now this manually add the depenency on test variants as needed.

Bug: 64939383
Change-Id: I2018caba9f5baf5622b39a2d38bce79a360e3f22
Test: updated RsSupportModeTest