Move the preNDK toolchain location
Currently, we store the preNDK toolchain for all abi's in a common
location. If this file is deleted, we'll end up recompiling for all
abi's and in some corner cases, we see "unable to find toolchain file"
error. Hence move the toolchain under the output folder for each abi so
it's handy for user to see and we'll not see compiler errors that are
hard to debug.

Test: 1. integration test
2. used it on a sample app (on both windows and linux) and verified that
the file is created within each abi's output folder.

Change-Id: I7601e67e6a4d702438d3fd1c366ffcb85800a395