Add timeout support for ReadWriteProcessLock.
In an upcoming CL, ReadWriteProcessLock will be used in the IDE code,
where it needs to be responsive to user events (such as cancellation).

This commit adds new methods in ReadWriteProcessLock and
ReadWriteThreadLock to acquire a lock with a timeout, so that those
methods return early if the lock cannot be acquired after the specified
time.

Bug: 64688589
Test: New tests in ReadWriteProcessLockTest, ReadWriteThreadLockTest
Change-Id: Ic73fa5c4c2fab2abf07a9a6a965fea97a96f873c