Work around Gradle bug with @Optional for @OutputFile/Dir.
If an @Optional @OutputFile/@OutputDirectory property is set to null in
a build, subsequent builds will always be UP-TO-DATE even if the output
file/directory is set to a different value.

This is a Gradle bug. See https://issuetracker.google.com/67418335 for
more details.

The workaround is to create a new property combining those properties
and make this new property an @Input. This way, the build will not be
UP-TO-DATE if any of the output files/directories changes.

Test: MergeResourcesTest, CombinedInputTest
Bug: 65829618, 67418335
Change-Id: Ic673c0734869fd158a6c90acd48a38d5dc9f4c57