Teach the tools/base performance tests how to upload their data to act-d.
The above is a small lie. At the moment they're not uploading to act-d,
instead they're uploading to a test instance that gcabillic@ has set up
for us to evaluate our usage of the service.

This does not disable the uploading to Google Cloud Storage (GCS) and
even if we decide to use act-d in the long term, I imagine we will still
want to keep uploading the data in to GCS. Having it available in Dremel
is useful, and I have no plans to remove that functionality in the near
or long term.

Change-Id: Ic48140af11f138e9afd43d5ae5c0631f71ea3365
Test: ./gradlew :base:integration-test:performanceTest